pyproject.toml
[tool.poetry]
name = "ev3-robonomics-client"
version = "0.1.0"
description = ""
authors = ["PaTara43 <p040399@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
robonomics-interface = "^1.3.0"
paho-mqtt = "^1.6.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
